+++ tests/test_cli.py
@@ -1,6 +1,7 @@
 # This file was part of Flask-CLI and was modified under the terms of
 # its Revised BSD License. Copyright © 2015 CERN.
 import os
+import platform
 import ssl
 import sys
 import types
@@ -17,6 +18,7 @@ from flask import Blueprint
 from flask import current_app
 from flask import Flask
 from flask.cli import AppGroup
+from flask.cli import DispatchingApp
 from flask.cli import dotenv
 from flask.cli import find_best_app
 from flask.cli import FlaskGroup
@@ -47,51 +49,36 @@ def test_cli_name(test_apps):
 
 
 def test_find_best_app(test_apps):
-    script_info = ScriptInfo()
-
     class Module:
         app = Flask("appname")
 
-    assert find_best_app(script_info, Module) == Module.app
+    assert find_best_app(Module) == Module.app
 
     class Module:
         application = Flask("appname")
 
-    assert find_best_app(script_info, Module) == Module.application
+    assert find_best_app(Module) == Module.application
 
     class Module:
         myapp = Flask("appname")
 
-    assert find_best_app(script_info, Module) == Module.myapp
+    assert find_best_app(Module) == Module.myapp
 
     class Module:
         @staticmethod
         def create_app():
             return Flask("appname")
 
-    app = find_best_app(script_info, Module)
+    app = find_best_app(Module)
     assert isinstance(app, Flask)
     assert app.name == "appname"
 
     class Module:
         @staticmethod
-        def create_app(foo):
+        def create_app(**kwargs):
             return Flask("appname")
 
-    with pytest.deprecated_call(match="Script info"):
-        app = find_best_app(script_info, Module)
-
-    assert isinstance(app, Flask)
-    assert app.name == "appname"
-
-    class Module:
-        @staticmethod
-        def create_app(foo=None, script_info=None):
-            return Flask("appname")
-
-    with pytest.deprecated_call(match="script_info"):
-        app = find_best_app(script_info, Module)
-
+    app = find_best_app(Module)
     assert isinstance(app, Flask)
     assert app.name == "appname"
 
@@ -100,7 +87,7 @@ def test_find_best_app(test_apps):
         def make_app():
             return Flask("appname")
 
-    app = find_best_app(script_info, Module)
+    app = find_best_app(Module)
     assert isinstance(app, Flask)
     assert app.name == "appname"
 
@@ -111,7 +98,7 @@ def test_find_best_app(test_apps):
         def create_app():
             return Flask("appname2")
 
-    assert find_best_app(script_info, Module) == Module.myapp
+    assert find_best_app(Module) == Module.myapp
 
     class Module:
         myapp = Flask("appname1")
@@ -120,32 +107,32 @@ def test_find_best_app(test_apps):
         def create_app():
             return Flask("appname2")
 
-    assert find_best_app(script_info, Module) == Module.myapp
+    assert find_best_app(Module) == Module.myapp
 
     class Module:
         pass
 
-    pytest.raises(NoAppException, find_best_app, script_info, Module)
+    pytest.raises(NoAppException, find_best_app, Module)
 
     class Module:
         myapp1 = Flask("appname1")
         myapp2 = Flask("appname2")
 
-    pytest.raises(NoAppException, find_best_app, script_info, Module)
+    pytest.raises(NoAppException, find_best_app, Module)
 
     class Module:
         @staticmethod
         def create_app(foo, bar):
             return Flask("appname2")
 
-    pytest.raises(NoAppException, find_best_app, script_info, Module)
+    pytest.raises(NoAppException, find_best_app, Module)
 
     class Module:
         @staticmethod
         def create_app():
             raise TypeError("bad bad factory!")
 
-    pytest.raises(TypeError, find_best_app, script_info, Module)
+    pytest.raises(TypeError, find_best_app, Module)
 
 
 @pytest.mark.parametrize(
@@ -209,8 +196,7 @@ def test_prepare_import(request, value, path, result):
     ),
 )
 def test_locate_app(test_apps, iname, aname, result):
-    info = ScriptInfo()
-    assert locate_app(info, iname, aname).name == result
+    assert locate_app(iname, aname).name == result
 
 
 @pytest.mark.parametrize(
@@ -232,20 +218,17 @@ def test_locate_app(test_apps, iname, aname, result):
     ),
 )
 def test_locate_app_raises(test_apps, iname, aname):
-    info = ScriptInfo()
-
     with pytest.raises(NoAppException):
-        locate_app(info, iname, aname)
+        locate_app(iname, aname)
 
 
 def test_locate_app_suppress_raise(test_apps):
-    info = ScriptInfo()
-    app = locate_app(info, "notanapp.py", None, raise_if_not_found=False)
+    app = locate_app("notanapp.py", None, raise_if_not_found=False)
     assert app is None
 
     # only direct import error is suppressed
     with pytest.raises(NoAppException):
-        locate_app(info, "cliapp.importerrorapp", None, raise_if_not_found=False)
+        locate_app("cliapp.importerrorapp", None, raise_if_not_found=False)
 
 
 def test_get_version(test_apps, capsys):
@@ -308,6 +291,26 @@ def test_scriptinfo(test_apps, monkeypatch):
     obj = ScriptInfo()
     app = obj.load_app()
     assert app.name == "testapp"
+
+
+@pytest.mark.xfail(platform.python_implementation() == "PyPy", reason="flaky on pypy")
+def test_lazy_load_error(monkeypatch):
+    """When using lazy loading, the correct exception should be
+    re-raised.
+    """
+
+    class BadExc(Exception):
+        pass
+
+    def bad_load():
+        raise BadExc
+
+    lazy = DispatchingApp(bad_load, use_eager_loading=False)
+
+    with pytest.raises(BadExc):
+        # reduce flakiness by waiting for the internal loading lock
+        with lazy._lock:
+            lazy._flush_bg_loading_exception()
 
 
 def test_with_appcontext(runner):